+++ actix-http/src/lib.rs
@@ -32,7 +32,7 @@ mod response;
 mod service;
 mod time_parser;
 
-pub mod cookie;
+pub use cookie;
 pub mod error;
 pub mod h1;
 pub mod h2;